requirements.txt
asgiref==3.4.1
Django==3.2.8
psycopg2==2.9.1
pytz==2021.3
sqlparse==0.4.2
djangorestframework~=3.12.4
celery~=5.2.1


